1. BaseCell을 만들어서 지정생성자 및 필수생성자를 기본으로 가지는 부모클래스를 만들어줌.
